+++ src/RoadCaptain.App.RouteBuilder/DummyUserPreferences.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Drawing;
 
 namespace RoadCaptain.App.RouteBuilder
 {
@@ -8,11 +7,12 @@ namespace RoadCaptain.App.RouteBuilder
         public string? DefaultSport { get; set; } = "Cycling";
         public string? LastUsedFolder { get; set; }
         public string? Route { get; set; }
-        public Point? InGameWindowLocation { get; set; } = new Point(600, 200);
+        public CapturedWindowLocation? InGameWindowLocation { get; set; } = new(600, 200, false, 800, 600);
         public bool EndActivityAtEndOfRoute { get; set; }
         public Version LastOpenedVersion { get; set; } = new Version(0, 0, 0, 0);
 
         public byte[]? ConnectionSecret { get; }
+        public CapturedWindowLocation? RouteBuilderLocation { get; set; }
 
         public void Load()
         {