+++ test/RoadCaptain.Tests.Unit/TestUserPreferences.cs
@@ -8,10 +8,12 @@ namespace RoadCaptain.Tests.Unit
         public string? DefaultSport { get; set; }
         public string? LastUsedFolder { get; set; }
         public string? Route { get; set; }
-        public Point? InGameWindowLocation { get; set; }
+        public CapturedWindowLocation? InGameWindowLocation { get; set; }
         public bool EndActivityAtEndOfRoute { get; set; }
         public Version LastOpenedVersion { get; set; }
         public byte[]? ConnectionSecret { get; set; }
+        public CapturedWindowLocation? RouteBuilderLocation { get; set; }
+
         public void Load()
         {
         }